requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
opencv-python==4.1.2.30
Pillow==7.2.0
pyzbar==0.1.8
Werkzeug==1.0.1
